backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.7
django-axes==6.5.1
django-cors-headers==4.4.0
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
phonenumbers==8.13.40
psycopg2-binary==2.9.9
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
